Feature: Segment list view
Added a segment list view to WindowsForms and Avalonia client. The segment list view can be displayed by using the 'View > Segment List` menu command. Future commits will make the list editable.

Added IWindowPaneFactory service that abstracts the creation of IWindowPanes, the way IDialogFactory abstracts the creation of IDialogs.

Changed Cmdids to an enum, to avoid having to manually renumber menu command ids

Generalized HexDisassemblerService and moved it into Reko.Core.Gui.

Created a stub OpenAsDialog for Avalonia.

+++ src/UnitTests/Gui/Design/ProgramDesignerTests.cs
@@ -55,7 +55,7 @@ namespace Reko.UnitTests.Gui.Design
             this.node = new Mock<ITreeNode>();
             this.host = new Mock<ITreeNodeDesignerHost>();
             this.uiSvc = new Mock<IDecompilerShellUiService>();
-            this.cmdidLoadSymbols = new CommandID(CmdSets.GuidReko, CmdIds.LoadSymbols);
+            this.cmdidLoadSymbols = new CommandID(CmdSets.GuidReko, (int)CmdIds.LoadSymbols);
             this.slSvc = new Mock<ISymbolLoadingService>();
             this.symSrc = new Mock<ISymbolSource>();
